IsometricPEPS

Documentation for IsometricPEPS.

IsometricPEPS.LMPS

IsometricPEPS.left_canonicalize!

IsometricPEPS.steifel_update!

IsometricPEPS.transverse_Ising_1d_effH

IsometricPEPS.transverse_Ising_1d_seperate_effH

IsometricPEPS.truncated_svd

IsometricPEPS.LMPS — Type

LMPS(left::AbstractTensor, tensors::Vector{AbstractTensor})

LMPS represents a Matrix Product State (MPS) with a left boundary tensor of order 2 (left) and a vector of tensors of order 3 (tensors).

Arguments

left::AbstractTensor: The left boundary tensor of order 2.

tensors::Vector{AbstractTensor}: The vector of tensors of order 3.

IsometricPEPS.left_canonicalize! — Function

left_canonicalize!(mps::LMPS, atol::Real=0)

Perform left canonicalization on an LMPS.

Arguments

mps::LMPS: The LMPS to be left canonicalized.

atol::Real=0: Truncation tolerance for performing SVD.

Output

An LMPS which consists of an isometric left boundary tensor of order 2 and (n-1) left canonical tensors of order 3.

IsometricPEPS.steifel_update! — Method

steifel_update!(mps::LMPS, G::@NamedTuple{left::Matrix{T}, tensors::Vector{Array{T, 3}}}, τ::T) where T

Update the LMPS mps using the gradient G and the updating rate τ. The gradient G should have the same shape as mps.

Arguments

mps::LMPS: The LMPS to update.

G::@NamedTuple{left::Matrix{T}, tensors::Vector{Array{T, 3}}}: The gradient of mps, which shares the same shape as mps.

τ::T: The updating rate.

Description

This function calculates the projected gradient on the tangent plane and updates mps with respect to the updating rate τ.

IsometricPEPS.transverse_Ising_1d_effH — Method

transverse_Ising_1d_effH(mps::LMPS, h::Float64, J::Float64)

This function generates the effective Hamiltonian for a 1d transverse Ising model by contracting ZZ and X terms together for every two sites.

Arguments

mps::LMPS: The LMPS representing the state of the system.

J::Float64: The coupling strength between adjacent spins.

h::Float64: The strength of the transverse magnetic field.

Returns

eff_H::Matrix{ComplexF64}: The effective Hamiltonian matrix.

IsometricPEPS.transverse_Ising_1d_seperate_effH — Method

transverse_Ising_1d_seperate_effH(mps::LMPS, h::Float64, J::Float64)

This function generates the effective Hamiltonian for a 1d transverse Ising model by contracting ZZ and X terms seperately.

Arguments

mps::LMPS: The LMPS representing the state of the system.

J::Float64: The coupling strength between adjacent spins.

h::Float64: The strength of the transverse magnetic field.

Returns

eff_H::Matrix{ComplexF64}: The effective Hamiltonian matrix.

IsometricPEPS.truncated_svd — Function

truncated_svd(tmat::AbstractMatrix, dmax::Int, atol::Real=0)

Perform singular value decomposition (SVD) for tmat and keep only the largest dmax singular values (not less than atol) and associated columns and rows.

Arguments

tmat::AbstractMatrix: The matrix to perform SVD on.

dmax::Int: The maximum number of singular values to keep.

atol::Real=0: The minimum threshold for singular values.

Returns

U: The left singular vectors.

S: The singular values.

Vt: The right singular vectors.

sum_discarded: The sum of discarded singular values.